Move plugin-related code to common module
