+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "matricks"
-version = "0.3.2-alpha.3"
+version = "0.3.2-alpha.4"
 edition = "2021"
 authors = ["Will McGloughlin <willem.mcg@gmail.com>"]
 license = "MIT"
