+++ src/plugin/plugin_iterator.rs
